data/requirements.txt
matplotlib
numpy
tqdm
Pillow
